// src/app/core/models/Hoghoghi/ShakhsHoghooghi.ts
export interface ShakhsHoghooghi {
    Id?: number;
    ShenasehMeli?: string;
    ShakhsHoghooghiName?: string;
    ShakhsHoghooghiTypeId?: number;
    ShakhsHoghooghiTypeName?: string;
    SabtDate?: any;
    ShomarehSabt?: string;
    MahalSabtCityId?: number;
    EghtesadiGhadimCode?: string;
    EghtesadiJadidCode?: string;
    AkharinRooznamehRasmiDate?: any;
    ShomarehMobile?: string;
    CityName?: string;
    CityId?: number
    ProvinceName?: string;
    ProvinceId?: any;
    ProvinceCode?: any
}